chore: reduced line code in cov page index
