+++ src/app/modules/signatory-data/submodules/coveragePage/index.tsx
@@ -2,7 +2,6 @@ import React, { useEffect } from 'react';
 import { withRouter } from 'react-router';
 
 import { CoverageLayout } from './layout';
-import moment from 'moment';
 
 /* store */
 import { covStore } from './store';
@@ -13,6 +12,7 @@ import { baseCovTable, covOrgQuery, covQuery } from './const';
 /* utils */
 import get from 'lodash/get';
 import { formatCovData } from './utils/formatCovData';
+import { getRangeStart } from './utils/getRangeStart';
 
 export function CoverageF(props) {
   const [state, actions] = covStore();
@@ -32,30 +32,14 @@ export function CoverageF(props) {
   useEffect(() => {
     const orgData = get(state.covOrg, 'data.data.response.docs[0]', null);
     if (orgData) {
-      let periodRange = perRange;
-      const totExps = get(orgData, 'organisation_total_expenditure', []);
-      let startDateTxt = '1900-01-01T00:00:00Z';
-      if (totExps.length > 0) {
-        const lastItem = totExps[totExps.length - 1];
-
-        const startDate = moment(lastItem.period_start);
-        let startMonth = startDate.month() + 1 + '';
-        startMonth = startMonth.length < 2 ? `0${startMonth}` : startMonth;
-        let startDay = startDate.date() + '';
-        startDay = startDay.length < 2 ? `0${startDay}` : startDay;
-        startDateTxt = `1900-${startMonth}-${startDay}T00:00:00Z`;
-        // we also add one day here so that moment would include the
-        // last day into the difference calculation
-        const endDate = moment(lastItem.period_end).add(1, 'days');
-        periodRange = endDate.diff(startDate, 'months');
-        setPerRange(periodRange);
-      }
+      const dateData = getRangeStart(orgData, perRange);
+      setPerRange(dateData.periodRange);
 
       actions.coverage.fetch({
         values: covQuery(
           decodeURIComponent(props.match.params.code),
-          periodRange,
-          startDateTxt
+          dateData.periodRange,
+          dateData.startDateTxt
         ),
       });
     }
